Rename remix to app
