+++ packages/react-sdk/src/app/root.tsx
