+++ packages/react-sdk/src/index.ts
@@ -5,4 +5,4 @@ export * from "./components/meta";
 export * from "./user-props";
 export * from "./pubsub";
 export * from "./db";
-export * from "./remix";
+export * from "./app";
