+++ packages/react-sdk/tsconfig.json
@@ -36,6 +36,10 @@
     },
 
     {
+      "path": "../image"
+    },
+
+    {
       "path": "../prisma-client"
     }
   ]
